DI_Study/DASH_JIRA/DASH_JIRA_2/DASH_JIRA_3_CODEX/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.5
pandas==2.2.3
langdetect==1.0.9
scikit-learn==1.4.2
openai==1.52.0
httpx==0.27.2
python-dotenv==1.0.1
